requirements.txt
Django
langchain
sentence-transformers
faiss-cpu
torch
transformers